# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/copropiedades/{id}/autorizaciones": {
      "get": {
        "operationId": "AutorizacionesController_listar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ver",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "activas",
                "historial"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorizacionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Autorizaciones activas o historial completo (HU-16, HU-17)",
        "tags": [
          "autorizaciones"
        ]
      },
      "post": {
        "operationId": "AutorizacionesController_crearAutorizacion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAutorizacionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdAutorizacionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea una autorización, única o recurrente (HU-07, HU-09, RN-22)",
        "tags": [
          "autorizaciones"
        ]
      }
    },
    "/copropiedades/{id}/autorizaciones/{autorizacionId}": {
      "put": {
        "operationId": "AutorizacionesController_modificarAutorizacion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autorizacionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarAutorizacionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModificacionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambia fin de vigencia, placa u observaciones de una viva (O3)",
        "tags": [
          "autorizaciones"
        ]
      }
    },
    "/copropiedades/{id}/autorizaciones/{autorizacionId}/acompanantes": {
      "post": {
        "operationId": "AutorizacionesController_agregarAcompanante",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autorizacionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarAcompananteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcompananteAgregadoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Añade un acompañante por su propia identidad (HU-08, D-01)",
        "tags": [
          "autorizaciones"
        ]
      }
    },
    "/copropiedades/{id}/autorizaciones/{autorizacionId}/fotografia": {
      "post": {
        "operationId": "AutorizacionesController_adjuntarFotografia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autorizacionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FotografiaDeVisitanteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FotografiaAdjuntadaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adjunta la fotografía de identificación del visitante (O3, RN-21)",
        "tags": [
          "autorizaciones"
        ]
      },
      "get": {
        "operationId": "AutorizacionesController_fotografia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autorizacionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlDeFotografiaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL firmada de vida corta de la fotografía del visitante (RN-21)",
        "tags": [
          "autorizaciones"
        ]
      }
    },
    "/copropiedades/{id}/autorizaciones/{autorizacionId}/revocacion": {
      "post": {
        "operationId": "AutorizacionesController_revocar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "autorizacionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevocarAutorizacionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocacionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoca con motivo obligatorio; no borra (HU-10, RN-19)",
        "tags": [
          "autorizaciones"
        ]
      }
    },
...
  },
  "info": {
    "title": "Next Control Residencial — API",
    "description": "Next Control decide. El hardware ejecuta.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AcompananteAgregadoDto": {
        "type": "object",
        "properties": {
          "agregado": {
            "type": "boolean"
          }
        },
        "required": [
          "agregado"
        ]
      },
      "AgregarAcompananteDto": {
        "type": "object",
        "properties": {
          "personaId": {
            "type": "string",
            "format": "uuid",
            "description": "El acompañante entra por su PROPIA identidad, para que la lista negra lo alcance (D-01)."
          },
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "personaId",
          "nombre"
        ]
      },
...
      "AutorizacionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "viviendaId": {
            "type": "string",
            "format": "uuid"
          },
          "vivienda": {
            "type": "string"
          },
          "visitante": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "desde": {
            "type": "string",
            "format": "date-time"
          },
          "hasta": {
            "type": "string",
            "format": "date-time"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "unica",
              "recurrente"
            ]
          },
          "estado": {
            "type": "string",
            "enum": [
              "activa",
              "revocada"
            ]
          },
          "placa": {
            "type": "string",
            "nullable": true
          },
          "acompanantes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "patron": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PatronDeAutorizacionDto"
              }
            ]
          },
          "revocadaEn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "motivoRevocacion": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "tieneFotografia": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "viviendaId",
          "vivienda",
          "visitante",
          "documento",
          "desde",
          "hasta",
          "tipo",
          "estado",
          "placa",
          "acompanantes",
          "patron",
          "revocadaEn",
          "motivoRevocacion",
          "observaciones",
          "tieneFotografia"
        ]
      },
...
      "CrearAutorizacionDto": {
        "type": "object",
        "properties": {
          "viviendaId": {
            "type": "string",
            "format": "uuid"
          },
          "personaId": {
            "type": "string",
            "format": "uuid",
            "description": "Persona que visita."
          },
          "desde": {
            "type": "string",
            "format": "date-time"
          },
          "hasta": {
            "type": "string",
            "format": "date-time"
          },
          "zonasPermitidas": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "maximoAcompanantes": {
            "type": "number",
            "minimum": 0,
            "maximum": 50
          },
          "patron": {
            "description": "Su presencia es lo único que distingue una recurrente de una única (HU-09).",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatronDeEntradaDto"
              }
            ]
          },
          "placa": {
            "type": "string",
            "nullable": true,
            "minLength": 5,
            "maxLength": 12
          },
          "observaciones": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          }
        },
        "required": [
          "viviendaId",
          "personaId",
          "desde",
          "hasta"
        ]
      },
...
      "FotografiaAdjuntadaDto": {
        "type": "object",
        "properties": {
          "adjuntada": {
            "type": "boolean"
          },
          "tipoMime": {
            "type": "string"
          },
          "tamanoBytes": {
            "type": "number"
          }
        },
        "required": [
          "adjuntada",
          "tipoMime",
          "tamanoBytes"
        ]
      },
      "FotografiaDeVisitanteDto": {
        "type": "object",
        "properties": {
          "tipoMime": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png"
            ]
          },
          "contenidoBase64": {
            "type": "string",
            "description": "Imagen JPEG o PNG en base64, máximo 1,5 MiB."
          }
        },
        "required": [
          "tipoMime",
          "contenidoBase64"
        ]
      },
...
      "IdAutorizacionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ModificacionDto": {
        "type": "object",
        "properties": {
          "modificada": {
            "type": "boolean"
          }
        },
        "required": [
          "modificada"
        ]
      },
      "ModificarAutorizacionDto": {
        "type": "object",
        "properties": {
          "hasta": {
            "type": "string",
            "format": "date-time"
          },
          "placa": {
            "type": "string",
            "nullable": true,
            "minLength": 5,
            "maxLength": 12
          },
          "observaciones": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000
          }
        }
      },
...
      "PatronDeAutorizacionDto": {
        "type": "object",
        "properties": {
          "dias": {
            "description": "Días de la semana, 0..6 con domingo = 0 (el vocabulario del dominio).",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "horaInicio": {
            "type": "string",
            "example": "08:00"
          },
          "horaFin": {
            "type": "string",
            "example": "18:00"
          }
        },
        "required": [
          "dias",
          "horaInicio",
          "horaFin"
        ]
      },
      "PatronDeEntradaDto": {
        "type": "object",
        "properties": {
          "dias": {
            "description": "Días de la semana, 0..6 con domingo = 0 (RN-22).",
            "example": [
              1,
              2,
              3,
              4,
              5
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "minutoInicio": {
            "type": "number",
            "minimum": 0,
            "maximum": 1440,
            "example": 480
          },
          "minutoFin": {
            "type": "number",
            "minimum": 0,
            "maximum": 1440,
            "example": 1080
          },
          "desplazamientoUtcMinutos": {
            "type": "number",
            "description": "Desfase UTC en minutos de la copropiedad; el patrón es local, no UTC.",
            "example": -300
          }
        },
        "required": [
          "dias",
          "minutoInicio",
          "minutoFin",
          "desplazamientoUtcMinutos"
        ]
      },
...
      "RevocacionDto": {
        "type": "object",
        "properties": {
          "revocada": {
            "type": "boolean"
          }
        },
        "required": [
          "revocada"
        ]
      },
      "RevocarAutorizacionDto": {
        "type": "object",
        "properties": {
          "motivo": {
            "type": "string",
            "minLength": 3,
            "maxLength": 300
          }
        },
        "required": [
          "motivo"
        ]
      },
...
      "UrlDeFotografiaDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "expiraEnSegundos": {
            "type": "number"
          }
        },
        "required": [
          "url",
          "expiraEnSegundos"
        ]
      },
...
    }
  }
}